# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Talos API
  description: >
    Source of truth for the talos-api public (/api/v1) and internal (/internal/v1) REST surface.
    The Angular client is generated from this file (openapi-generator, typescript-angular); CI
    diffs the running API's springdoc output against it. Endpoints are added phase by phase per
    Section 10 of the implementation plan — this file currently covers Phase 1 (auth),
    Phase 3 (project registry), Phase 4 (task board), Phase 5 (agent run lifecycle, API side),
    Phase 6 (cancel + internal changes ingest), Phase 8 (review/approval flow: diff serving and
    the approvals endpoints), Phase 9 (git push and PR workflow: integrations, the
    git-token/pull-request internal endpoints, and GET /runs/{id}/pull-request), Phase 10
    (Dokploy integration), Phase 12 (chat trigger support), Phase 13 (memory ingestion and
    internal retrieval), Phase 14 (per-run cost tracking and advisory recommendations), and
    Phase 15 (multi-user RBAC enforcement and user management).
  version: "0.15.0"
servers:
  - url: /api/v1
    description: "Public API (JWT via Authorization: Bearer)"
  - url: /internal/v1
    description: "Internal API (service token via X-Talos-Internal-Token)"
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      operationId: listTasks
      summary: List tasks.
      tags: [tasks]
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TaskStatus"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageTaskSummary"
    post:
      operationId: createTask
      summary: Create a task in BACKLOG.
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /tasks/{id}:
    get:
      operationId: getTask
      summary: Get a task and its agent runs.
      tags: [tasks]
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetail"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      operationId: patchTask
      summary: Partially update task fields (not status — use /move).
      tags: [tasks]
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTaskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /tasks/{id}/move:
    post:
      operationId: moveTask
      summary: Move a task to a new Kanban column and/or reorder it within one.
      tags: [tasks]
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveTaskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          description: Illegal status transition (Section 8.2 -- RUNNING/REVIEW/DONE/BLOCKED are set only by the run state machine).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{id}/start-run:
    post:
      operationId: startRun
      summary: Start an agent run for a task (CREATED -> QUEUED, publishes task.run.requested).
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRunRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          description: An active (non-terminal) run already exists for this task.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceToken:
      type: apiKey
      in: header
      name: X-Talos-Internal-Token
  parameters:
...
    TaskId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error envelope (Section 10.1).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      description: Error envelope returned on every non-2xx response (Section 10.1).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: SCREAMING_SNAKE error code enumerated in this spec.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    RunSummary:
      type: object
      required: [id, taskId, status, agentKey, createdAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        status:
          type: string
        agentKey:
          type: string
        createdAt:
          type: string
          format: date-time
...
    TaskStatus:
      type: string
      enum: [BACKLOG, READY, RUNNING, REVIEW, BLOCKED, DONE, CANCELLED]
    TaskPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    TaskRiskLevel:
      type: string
      enum: [NORMAL, HIGH]
    TaskSummary:
      type: object
      required: [id, projectId, title, source, status, priority, riskLevel, boardPosition, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        source:
          type: string
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        riskLevel:
          $ref: "#/components/schemas/TaskRiskLevel"
        boardPosition:
          type: integer
        requestedBy:
          type: string
          format: uuid
        assignedAgentKey:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Task:
      allOf:
        - $ref: "#/components/schemas/TaskSummary"
    TaskDetail:
      allOf:
        - $ref: "#/components/schemas/TaskSummary"
        - type: object
          required: [runs]
          properties:
            runs:
              type: array
              items:
                $ref: "#/components/schemas/RunSummary"
    CreateTaskRequest:
      type: object
      required: [projectId, title]
      properties:
        projectId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        riskLevel:
          $ref: "#/components/schemas/TaskRiskLevel"
        source:
          type: string
          description: >-
            Provenance, e.g. DASHBOARD (default) | WEBHOOK | TELEGRAM | WHATSAPP (Section 16 Phase
            12 Track B). Unrecognized values fall back to DASHBOARD.
        assignedAgentKey:
          type: string
          description: >-
            Phase 14: sets the task's preferred agent at creation time, e.g. following a
            recommendations hint -- an explicit operator choice, never auto-selected. Used by
            start-run as a fallback when the run request itself doesn't specify agentKey.
    PatchTaskRequest:
      type: object
      description: Only non-null fields are applied. Status is not patchable here -- use /move.
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          $ref: "#/components/schemas/TaskPriority"
        riskLevel:
          $ref: "#/components/schemas/TaskRiskLevel"
        assignedAgentKey:
          type: string
    MoveTaskRequest:
      type: object
      required: [status, boardPosition]
      properties:
        status:
          $ref: "#/components/schemas/TaskStatus"
        boardPosition:
          type: integer
    PageTaskSummary:
      type: object
      required: [content, page, size, totalElements]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
    RunStatus:
      type: string
      enum: [CREATED, QUEUED, PREPARING_WORKSPACE, RUNNING_AGENT, RUNNING_TESTS, REVIEWING,
        WAITING_APPROVAL, APPROVED, REJECTED, COMPLETED, FAILED, CANCELLED]
    TestStatus:
      type: string
      enum: [NOT_RUN, PASSED, FAILED, ERROR]
    ReviewStatus:
      type: string
      enum: [CLEAN, RISK_FLAGGED]
...
    Run:
      type: object
      required: [id, taskId, projectId, status, agentKey, providerAuthMode, testStatus, reviewStatus,
        createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        agentKey:
          type: string
        providerAuthMode:
          type: string
        prompt:
          type: string
        branchName:
          type: string
        workspacePath:
          type: string
        summary:
          type: string
        testStatus:
          $ref: "#/components/schemas/TestStatus"
        reviewStatus:
          $ref: "#/components/schemas/ReviewStatus"
        errorMessage:
          type: string
        exitCode:
          type: integer
        inputTokens:
          type: integer
          description: Phase 14 (Section 16) normalized usage metadata; null when the adapter can't report it.
        outputTokens:
          type: integer
        costUsd:
          type: number
          description: >-
            Provider-reported price in USD. Always null for provider_auth_mode=subscription_local
            (Section 13) -- never a fabricated/estimated price.
        costModel:
          type: string
        timeoutAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    StartRunRequest:
      type: object
      properties:
        agentKey:
          type: string
          description: Defaults to the active talos.yaml's agents.preferred when omitted.
        authMode:
          type: string
          description: Defaults to "api_key" when omitted.
...